% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ismb2012}
\alias{ismb2012}
\title{ISMB tweets from 2012}
\format{A data frame with 3162 rows and 5 variables}
\source{
\url{https://github.com/neilfws/Twitter/tree/master/ismb/data}
}
\description{
A set of tweets over 5 days with the hashtag #ISMB from 2012. The variables are:
}
\details{
\itemize{
  \item text. the text of the tweet
  \item created. the date and time of the tweet
  \item id. the twitter id of the tweet
  \item screenName. who sent the tweet
  \item hashSearch. the hashtag that was searched
}
}

