% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readTweetData}
\alias{readTweetData}
\title{read and process ST tweet data}
\usage{
readTweetData(file, hashSearch, widths = c(18, 14, 18, 1000))
}
\arguments{
\item{file}{the file to process}

\item{widths}{the widths of each field. Defaults based on ST's originals}
}
\value{
data.frame
}
\description{
reads in Stephen Turners archived twitter data and reformats it into a nicer format
}

