% Generated by roxygen2 (4.0.0): do not edit by hand
\name{retweetCount}
\alias{retweetCount}
\title{find rt counts}
\usage{
retweetCount(tweetDF)
}
\arguments{
\item{tweetDF}{the data.frame of tweets}
}
\value{
data.frame of original tweets with countRT added
}
\description{
for each original tweet, find the number of times it was retweeted.
}

