% Generated by roxygen2 (4.0.0): do not edit by hand
\name{totalRT}
\alias{totalRT}
\title{total retweeted}
\usage{
totalRT(tweetDF, countColumn)
}
\arguments{
\item{tweetDF}{data.frame of tweets}

\item{countColumn}{which column has the retweet counts (should be a text id)}
}
\value{
data.frame of counts
}
\description{
how often a users tweets were retweeted
}

