% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tweetRank}
\alias{tweetRank}
\title{total tweet rank}
\usage{
tweetRank(tweetCounts)
}
\arguments{
\item{tweetCounts}{a vector of tweets by user}
}
\value{
ranks by percentage of total from tweetCounts
}
\description{
calculates rank of a user by the percentage of total tweets
}

